refactor(mmio): move MMIODeviceManager into separate module
MMIODeviceManager's implementation was spread across 2 modules:
mmio.rs and persist.rs. By moving all implementation into mmio
module we wil have an easir time when it comes to add more
device managers (e.g. ACPI device_manager).
